Fix the leading space in the anchor title may not be trimmed
