+++ src/element_handler/anchor.rs
@@ -83,10 +83,9 @@ impl AnchorElementHandler {
 
     fn build_inlined_anchor(&self, content: &str, link: String, title: Option<String>) -> String {
         let has_spaces_in_link = link.contains(' ');
-        let (content, leading_whitespace) = content.strip_leading_whitespace();
+        let (content, _) = content.strip_leading_whitespace();
         let (content, trailing_whitespace) = content.strip_trailing_whitespace();
         concat_strings!(
-            leading_whitespace.unwrap_or(""),
             "[",
             content,
             "](",
