+++ tests/basic_tests.rs
@@ -6,17 +6,6 @@ use htmd::{
     Element, HtmlToMarkdown,
 };
 
-#[test]
-fn links() {
-    let html = r#"
-        <a href="https://example.com">Link 1</a>
-        <a href="https://example.com" title="Hello">Link 2</a>
-        "#;
-    assert_eq!(
-        "[Link 1](https://example.com)[Link 2](https://example.com \"Hello\")",
-        convert(html).unwrap(),
-    )
-}
 
 #[test]
 fn links_with_spaces() {
